api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic-settings==2.3.1

# Extracción de documentos
PyMuPDF==1.24.5
pytesseract==0.3.10
Pillow==10.3.0

# Embeddings
sentence-transformers==3.0.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu

# Qdrant
qdrant-client==1.10.1
